fix: child index on retry not being reset correctly
